Changes to reduce MSVC warnings
git-svn-id: http://svn.boost.org/svn/boost/trunk@20281 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/date_time/time_parsing.hpp
@@ -131,12 +131,12 @@ namespace date_time {
         }
       case 1: 
         {
-          min = boost::lexical_cast<int>(*ti); 
+          min = boost::lexical_cast<short>(*ti); 
           break;
         }
       case 2: 
        {
-         sec = boost::lexical_cast<int>(*ti); 
+         sec = boost::lexical_cast<short>(*ti); 
          break;
         }
       };
